refactor: datalake into primitives
