+++ cli/Cargo.toml
@@ -27,6 +27,7 @@ tokio = { workspace = true }
 serde_json = { workspace = true }
 tracing = { workspace = true }
 hdp-provider = { workspace = true }
+hdp-primitives = { workspace = true }
 
 clap = { version = "4.4.4", features = ["derive"] }
 dotenv = "0.15.0"
