+++ crates/core/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod codec;
 pub mod compiler;
 pub mod config;
-pub mod datalake;
 pub mod evaluator;
 pub mod task;
