+++ crates/primitives/Cargo.toml
@@ -9,5 +9,6 @@ version.workspace = true
 [dependencies]
 serde = { workspace = true, features = ["derive"] }
 anyhow = { workspace = true }
+alloy-dyn-abi = { workspace = true }
 alloy-rlp = { workspace = true, features = ["derive"] }
 alloy-primitives = { workspace = true, features = ["rlp", "serde"] }
