+++ crates/primitives/src/block/header.rs
@@ -1,4 +1,3 @@
-use anyhow::{bail, Result};
 use std::str::FromStr;
 
 use alloy_primitives::{hex, keccak256, Address, BlockNumber, Bloom, Bytes, B256, B64, U256};
@@ -277,13 +276,7 @@ impl Decodable for Header {
 
 // =============================================================================
 
-pub struct BlockHeader(Header);
-
-impl BlockHeader {
-    pub fn new_from_header(value: Header) -> Self {
-        BlockHeader(value)
-    }
-
+impl Header {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         parent_hash: B256,
@@ -307,7 +300,7 @@ impl BlockHeader {
         excess_blob_gas: Option<u64>,
         parent_beacon_block_root: Option<B256>,
     ) -> Self {
-        BlockHeader(Header {
+        Header {
             parent_hash,
             ommers_hash,
             beneficiary,
@@ -328,182 +321,21 @@ impl BlockHeader {
             blob_gas_used,
             excess_blob_gas,
             parent_beacon_block_root,
-        })
+        }
     }
 
     pub fn rlp_encode(&self) -> String {
         let mut buffer = Vec::<u8>::new();
-        self.0.encode(&mut buffer);
+        self.encode(&mut buffer);
         hex::encode(buffer)
     }
 
     pub fn rlp_decode(rlp: &str) -> Self {
-        let decoded = <Header>::decode(&mut hex::decode(rlp).unwrap().as_slice()).unwrap();
-        BlockHeader::new_from_header(decoded)
+        <Header>::decode(&mut hex::decode(rlp).unwrap().as_slice()).unwrap()
     }
 
     pub fn get_block_hash(&self) -> String {
-        self.0.hash_slow().to_string()
-    }
-}
-
-#[derive(Debug)]
-pub enum HeaderField {
-    ParentHash,
-    OmmerHash,
-    Beneficiary,
-    StateRoot,
-    TransactionsRoot,
-    ReceiptsRoot,
-    LogsBloom,
-    Difficulty,
-    Number,
-    GasLimit,
-    GasUsed,
-    Timestamp,
-    ExtraData,
-    MixHash,
-    Nonce,
-    BaseFeePerGas,
-    WithdrawalsRoot,
-    BlobGasUsed,
-    ExcessBlobGas,
-    ParentBeaconBlockRoot,
-}
-
-impl HeaderField {
-    pub fn from_index(index: u8) -> Option<Self> {
-        match index {
-            0 => Some(HeaderField::ParentHash),
-            1 => Some(HeaderField::OmmerHash),
-            2 => Some(HeaderField::Beneficiary),
-            3 => Some(HeaderField::StateRoot),
-            4 => Some(HeaderField::TransactionsRoot),
-            5 => Some(HeaderField::ReceiptsRoot),
-            6 => Some(HeaderField::LogsBloom),
-            7 => Some(HeaderField::Difficulty),
-            8 => Some(HeaderField::Number),
-            9 => Some(HeaderField::GasLimit),
-            10 => Some(HeaderField::GasUsed),
-            11 => Some(HeaderField::Timestamp),
-            12 => Some(HeaderField::ExtraData),
-            13 => Some(HeaderField::MixHash),
-            14 => Some(HeaderField::Nonce),
-            15 => Some(HeaderField::BaseFeePerGas),
-            16 => Some(HeaderField::WithdrawalsRoot),
-            17 => Some(HeaderField::BlobGasUsed),
-            18 => Some(HeaderField::ExcessBlobGas),
-            19 => Some(HeaderField::ParentBeaconBlockRoot),
-            _ => None,
-        }
-    }
-
-    pub fn to_index(&self) -> u8 {
-        match self {
-            HeaderField::ParentHash => 0,
-            HeaderField::OmmerHash => 1,
-            HeaderField::Beneficiary => 2,
-            HeaderField::StateRoot => 3,
-            HeaderField::TransactionsRoot => 4,
-            HeaderField::ReceiptsRoot => 5,
-            HeaderField::LogsBloom => 6,
-            HeaderField::Difficulty => 7,
-            HeaderField::Number => 8,
-            HeaderField::GasLimit => 9,
-            HeaderField::GasUsed => 10,
-            HeaderField::Timestamp => 11,
-            HeaderField::ExtraData => 12,
-            HeaderField::MixHash => 13,
-            HeaderField::Nonce => 14,
-            HeaderField::BaseFeePerGas => 15,
-            HeaderField::WithdrawalsRoot => 16,
-            HeaderField::BlobGasUsed => 17,
-            HeaderField::ExcessBlobGas => 18,
-            HeaderField::ParentBeaconBlockRoot => 19,
-        }
-    }
-
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            HeaderField::ParentHash => "PARENT_HASH",
-            HeaderField::OmmerHash => "OMMERS_HASH",
-            HeaderField::Beneficiary => "BENEFICIARY",
-            HeaderField::StateRoot => "STATE_ROOT",
-            HeaderField::TransactionsRoot => "TRANSACTIONS_ROOT",
-            HeaderField::ReceiptsRoot => "RECEIPTS_ROOT",
-            HeaderField::LogsBloom => "LOGS_BLOOM",
-            HeaderField::Difficulty => "DIFFICULTY",
-            HeaderField::Number => "NUMBER",
-            HeaderField::GasLimit => "GAS_LIMIT",
-            HeaderField::GasUsed => "GAS_USED",
-            HeaderField::Timestamp => "TIMESTAMP",
-            HeaderField::ExtraData => "EXTRA_DATA",
-            HeaderField::MixHash => "MIX_HASH",
-            HeaderField::Nonce => "NONCE",
-            HeaderField::BaseFeePerGas => "BASE_FEE_PER_GAS",
-            HeaderField::WithdrawalsRoot => "WITHDRAWALS_ROOT",
-            HeaderField::BlobGasUsed => "BLOB_GAS_USED",
-            HeaderField::ExcessBlobGas => "EXCESS_BLOB_GAS",
-            HeaderField::ParentBeaconBlockRoot => "PARENT_BEACON_BLOCK_ROOT",
-        }
-    }
-}
-
-impl FromStr for HeaderField {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> Result<Self> {
-        match s {
-            "PARENT_HASH" => Ok(HeaderField::ParentHash),
-            "OMMERS_HASH" => Ok(HeaderField::OmmerHash),
-            "BENEFICIARY" => Ok(HeaderField::Beneficiary),
-            "STATE_ROOT" => Ok(HeaderField::StateRoot),
-            "TRANSACTIONS_ROOT" => Ok(HeaderField::TransactionsRoot),
-            "RECEIPTS_ROOT" => Ok(HeaderField::ReceiptsRoot),
-            "LOGS_BLOOM" => Ok(HeaderField::LogsBloom),
-            "DIFFICULTY" => Ok(HeaderField::Difficulty),
-            "NUMBER" => Ok(HeaderField::Number),
-            "GAS_LIMIT" => Ok(HeaderField::GasLimit),
-            "GAS_USED" => Ok(HeaderField::GasUsed),
-            "TIMESTAMP" => Ok(HeaderField::Timestamp),
-            "EXTRA_DATA" => Ok(HeaderField::ExtraData),
-            "MIX_HASH" => Ok(HeaderField::MixHash),
-            "NONCE" => Ok(HeaderField::Nonce),
-            "BASE_FEE_PER_GAS" => Ok(HeaderField::BaseFeePerGas),
-            "WITHDRAWALS_ROOT" => Ok(HeaderField::WithdrawalsRoot),
-            "BLOB_GAS_USED" => Ok(HeaderField::BlobGasUsed),
-            "EXCESS_BLOB_GAS" => Ok(HeaderField::ExcessBlobGas),
-            "PARENT_BEACON_BLOCK_ROOT" => Ok(HeaderField::ParentBeaconBlockRoot),
-            _ => bail!("Unknown header field"),
-        }
-    }
-}
-
-pub fn decode_header_field(header_rlp: &str, field: HeaderField) -> String {
-    let decoded =
-        <Header as Decodable>::decode(&mut hex::decode(header_rlp).unwrap().as_slice()).unwrap();
-
-    match field {
-        HeaderField::ParentHash => decoded.parent_hash.to_string(),
-        HeaderField::OmmerHash => decoded.ommers_hash.to_string(),
-        HeaderField::Beneficiary => decoded.beneficiary.to_string(),
-        HeaderField::StateRoot => decoded.state_root.to_string(),
-        HeaderField::TransactionsRoot => decoded.transactions_root.to_string(),
-        HeaderField::ReceiptsRoot => decoded.receipts_root.to_string(),
-        HeaderField::LogsBloom => decoded.logs_bloom.to_string(),
-        HeaderField::Difficulty => decoded.difficulty.to_string(),
-        HeaderField::Number => decoded.number.to_string(),
-        HeaderField::GasLimit => decoded.gas_limit.to_string(),
-        HeaderField::GasUsed => decoded.gas_used.to_string(),
-        HeaderField::Timestamp => decoded.timestamp.to_string(),
-        HeaderField::ExtraData => decoded.extra_data.to_string(),
-        HeaderField::MixHash => decoded.mix_hash.to_string(),
-        HeaderField::Nonce => decoded.nonce.to_string(),
-        HeaderField::BaseFeePerGas => decoded.base_fee_per_gas.unwrap().to_string(),
-        HeaderField::WithdrawalsRoot => decoded.withdrawals_root.unwrap().to_string(),
-        HeaderField::BlobGasUsed => decoded.blob_gas_used.unwrap().to_string(),
-        HeaderField::ExcessBlobGas => decoded.excess_blob_gas.unwrap().to_string(),
-        HeaderField::ParentBeaconBlockRoot => decoded.parent_beacon_block_root.unwrap().to_string(),
+        self.hash_slow().to_string()
     }
 }
 
@@ -543,9 +375,9 @@ impl BlockHeaderFromRpc {
     }
 }
 
-impl From<&BlockHeaderFromRpc> for BlockHeader {
+impl From<&BlockHeaderFromRpc> for Header {
     fn from(value: &BlockHeaderFromRpc) -> Self {
-        Self(Header {
+        Self {
             parent_hash: B256::from_str(&value.parent_hash).expect("Invalid hex string"),
             ommers_hash: B256::from_str(&value.sha3_uncles).expect("Invalid hex string"),
             beneficiary: Address::from_str(&value.miner).expect("Invalid hex string"),
@@ -583,7 +415,7 @@ impl From<&BlockHeaderFromRpc> for BlockHeader {
                 .parent_beacon_block_root
                 .clone()
                 .map(|x| B256::from_str(&x).expect("Invalid hex string")),
-        })
+        }
     }
 }
