+++ crates/primitives/src/block/mod.rs
@@ -1,3 +1,2 @@
 pub mod account;
 pub mod header;
-pub mod transaction;
