+++ crates/primitives/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod block;
+pub mod datalake;
 pub mod format;
 pub mod utils;
